While giving print, the paper width is only 80mm.
and do not print the excess information outside the content. it should start with the company name and end with the thank you. leave the empty white spaces above it and below it. keep the eidth to 80mm. and try to limt the hieght by leaving the empty white spaces abve thebill and below the bill